[fix]scoreモデルとquestionモデルにjobの処理を分散した
